# api/specs/swagger.yaml
openapi: 3.0.0
info:
  title: Purchases API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
servers:
  - url: http://localhost:8888/purchases
    description: Optional server description, e.g. Main (production) server
paths:
...
  /auth/login:
    post:
      operationId: loginUser
      summary: "Logs user into the system via username amd password"
      tags:
        - Auth
      requestBody:
        description: "The user object for login"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAuth"
      responses:
        '200':
          description: "Received logged in user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /purchase:
    get:
      tags:
        - Purchase
      operationId: getPurchases
      summary: "Get list of all Purchases"
      parameters:
        - in: query
          required: false
          name: showAll
          description: "Id of the Purchase"
          schema:
            type: boolean
      responses:
        '200':
          description: "Retrieved list of all Purchases"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      tags:
        - Purchase
      operationId: createPurchase
      summary: "Creates new Purchase"
      requestBody:
        description: "Purchase"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Purchase"
      responses:
        '200':
          description: "Created Purchase"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
    put:
      tags:
        - Purchase
      operationId: editPurchase
      summary: "Edits Purchase"
      requestBody:
        description: "Purchase"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Purchase"
      responses:
        '200':
          description: "Edited Purchase"
  /purchase/{purchaseId}:
    delete:
      tags:
        - Purchase
      operationId: deletePurchase
      summary: "Deletes existing Purchase"
      parameters:
        - in: path
          required: true
          name: purchaseId
          description: "Id of the Purchase"
          schema:
            type: integer
      responses:
        '200':
          description: "Deleted Purchase"
  /purchase/getPurchaseById:
    get:
      tags:
        - Purchase
      operationId: getPurchaseById
      summary: "Get Purchase by id"
      parameters:
        - in: query
          required: true
          name: purchaseId
          description: "Id of the Purchase"
          schema:
            type: integer
        - in: query
          required: false
          name: showAll
          description: "Id of the Purchase"
          schema:
            type: boolean
      responses:
        '200':
          description: "Retrieved Purchase"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchase/fromTemplate:
    post:
      tags:
        - Purchase
      operationId: createFromTemplate
      summary: "creating Purchase from PurchaseTemplate"
      requestBody:
        description: "PurchaseTemplate"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseTemplate"
      responses:
        '200':
          description: "Retrieved Purchase"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchaseItem/toggle/{purchaseItemId}:
    put:
      tags:
        - Purchase Item
      operationId: togglePurchaseItem
      summary: "Toggle purchase item check"
      parameters:
        - in: path
          required: true
          name: purchaseItemId
          description: "Id of the PurchaseItem"
          schema:
            type: integer
      responses:
        '200':
          description: "Edited Purchase"
...
components:
  schemas:
    Purchase:
      type: object
      properties:
        id:
          type: integer
        createdDate:
          type: string
          format: date-time
        checked:
          type: boolean
        checkedDate:
          type: string
          format: date-time
        name:
          type: string
        text:
          type: string
        username:
          type: string
        shared:
          type: boolean
        sharedForUsername:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
    PurchaseItem:
      type: object
      properties:
        id:
          type: integer
        purchaseId:
          type: integer
        checked:
          type: boolean
        checkedDate:
          type: string
          format: date-time
        itemName:
          type: string
        itemDescription:
          type: string
        active:
          type: boolean
        ord:
          type: integer
    PurchaseTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        text:
          type: string
        username:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseTemplateItem"
    PurchaseTemplateItem:
      type: object
      properties:
        id:
          type: integer
        purchaseTemplateId:
          type: integer
        itemName:
          type: string
        itemDescription:
          type: string
        ord:
          type: integer
    UserAuth:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    User:
      type: object
      properties:
        username:
          type: string
        fio:
          type: string
        apiKey:
          type: string
        gender:
          type: string
...