# server/src/main/java/com/michael200kg/purchaseserver/openapi/api/PurchaseApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.2.2).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.michael200kg.purchaseserver.openapi.api;

import com.michael200kg.purchaseserver.openapi.dto.Purchase;
import com.michael200kg.purchaseserver.openapi.dto.PurchaseTemplate;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-01-12T13:45:49.110733200+03:00[Europe/Moscow]")

@Validated
@Api(value = "Purchase", description = "the Purchase API")
public interface PurchaseApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    @ApiOperation(value = "creating Purchase from PurchaseTemplate", nickname = "createFromTemplate", notes = "", response = Purchase.class, tags={ "Purchase", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Retrieved Purchase", response = Purchase.class) })
    @RequestMapping(value = "/purchase/fromTemplate",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Purchase> createFromTemplate(@ApiParam(value = "PurchaseTemplate"  )  @Valid @RequestBody PurchaseTemplate purchaseTemplate) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"shared\" : true, \"createdDate\" : \"2000-01-23T04:56:07.000+00:00\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"sharedForUsername\" : \"sharedForUsername\", \"name\" : \"name\", \"checked\" : true, \"id\" : 0, \"text\" : \"text\", \"items\" : [ { \"ord\" : 5, \"itemName\" : \"itemName\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"purchaseId\" : 1, \"checked\" : true, \"active\" : true, \"id\" : 6, \"itemDescription\" : \"itemDescription\" }, { \"ord\" : 5, \"itemName\" : \"itemName\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"purchaseId\" : 1, \"checked\" : true, \"active\" : true, \"id\" : 6, \"itemDescription\" : \"itemDescription\" } ], \"username\" : \"username\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Creates new Purchase", nickname = "createPurchase", notes = "", response = Purchase.class, tags={ "Purchase", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Created Purchase", response = Purchase.class) })
    @RequestMapping(value = "/purchase",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Purchase> createPurchase(@ApiParam(value = "Purchase"  )  @Valid @RequestBody Purchase purchase) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"shared\" : true, \"createdDate\" : \"2000-01-23T04:56:07.000+00:00\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"sharedForUsername\" : \"sharedForUsername\", \"name\" : \"name\", \"checked\" : true, \"id\" : 0, \"text\" : \"text\", \"items\" : [ { \"ord\" : 5, \"itemName\" : \"itemName\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"purchaseId\" : 1, \"checked\" : true, \"active\" : true, \"id\" : 6, \"itemDescription\" : \"itemDescription\" }, { \"ord\" : 5, \"itemName\" : \"itemName\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"purchaseId\" : 1, \"checked\" : true, \"active\" : true, \"id\" : 6, \"itemDescription\" : \"itemDescription\" } ], \"username\" : \"username\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Deletes existing Purchase", nickname = "deletePurchase", notes = "", tags={ "Purchase", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Deleted Purchase") })
    @RequestMapping(value = "/purchase/{purchaseId}",
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> deletePurchase(@ApiParam(value = "Id of the Purchase",required=true) @PathVariable("purchaseId") Integer purchaseId) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Edits Purchase", nickname = "editPurchase", notes = "", tags={ "Purchase", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Edited Purchase") })
    @RequestMapping(value = "/purchase",
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    default ResponseEntity<Void> editPurchase(@ApiParam(value = "Purchase"  )  @Valid @RequestBody Purchase purchase) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Get Purchase by id", nickname = "getPurchaseById", notes = "", response = Purchase.class, tags={ "Purchase", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Retrieved Purchase", response = Purchase.class) })
    @RequestMapping(value = "/purchase/getPurchaseById",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Purchase> getPurchaseById(@NotNull @ApiParam(value = "Id of the Purchase", required = true) @Valid @RequestParam(value = "purchaseId", required = true) Integer purchaseId,@ApiParam(value = "Id of the Purchase") @Valid @RequestParam(value = "showAll", required = false) Boolean showAll) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"shared\" : true, \"createdDate\" : \"2000-01-23T04:56:07.000+00:00\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"sharedForUsername\" : \"sharedForUsername\", \"name\" : \"name\", \"checked\" : true, \"id\" : 0, \"text\" : \"text\", \"items\" : [ { \"ord\" : 5, \"itemName\" : \"itemName\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"purchaseId\" : 1, \"checked\" : true, \"active\" : true, \"id\" : 6, \"itemDescription\" : \"itemDescription\" }, { \"ord\" : 5, \"itemName\" : \"itemName\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"purchaseId\" : 1, \"checked\" : true, \"active\" : true, \"id\" : 6, \"itemDescription\" : \"itemDescription\" } ], \"username\" : \"username\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Get list of all Purchases", nickname = "getPurchases", notes = "", response = Purchase.class, responseContainer = "List", tags={ "Purchase", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Retrieved list of all Purchases", response = Purchase.class, responseContainer = "List") })
    @RequestMapping(value = "/purchase",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Purchase>> getPurchases(@ApiParam(value = "Id of the Purchase") @Valid @RequestParam(value = "showAll", required = false) Boolean showAll) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"shared\" : true, \"createdDate\" : \"2000-01-23T04:56:07.000+00:00\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"sharedForUsername\" : \"sharedForUsername\", \"name\" : \"name\", \"checked\" : true, \"id\" : 0, \"text\" : \"text\", \"items\" : [ { \"ord\" : 5, \"itemName\" : \"itemName\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"purchaseId\" : 1, \"checked\" : true, \"active\" : true, \"id\" : 6, \"itemDescription\" : \"itemDescription\" }, { \"ord\" : 5, \"itemName\" : \"itemName\", \"checkedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"purchaseId\" : 1, \"checked\" : true, \"active\" : true, \"id\" : 6, \"itemDescription\" : \"itemDescription\" } ], \"username\" : \"username\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}